+++ adagucserverEC/CAreaMapper.cpp
@@ -30,13 +30,13 @@
  */
 
 #include "CAreaMapper.h"
-#include "CImageDataWriter.h"
+#include "CDrawFunction.h"
 
 const char *CAreaMapper::className = "CAreaMapper";
 
 void CAreaMapper::init(CDataSource *dataSource, CDrawImage *drawImage, int tileWidth, int tileHeight) {
   this->dataSource = dataSource;
-  this->drawImage = drawImage;
+  this->settings.drawImage = drawImage;
   dfTileWidth = double(tileWidth);
   dfTileHeight = double(tileHeight);
   for (int k = 0; k < 4; k++) {
@@ -65,21 +65,21 @@ void CAreaMapper::init(CDataSource *dataSource, CDrawImage *drawImage, int tileW
 
   CStyleConfiguration *styleConfiguration = dataSource->getStyle();
 
-  dfNodataValue = dataSource->getDataObject(0)->dfNodataValue;
-  legendValueRange = styleConfiguration->hasLegendValueRange;
-  legendLowerRange = styleConfiguration->legendLowerRange;
-  legendUpperRange = styleConfiguration->legendUpperRange;
-  hasNodataValue = dataSource->getDataObject(0)->hasNodataValue;
+  settings.dfNodataValue = dataSource->getDataObject(0)->dfNodataValue;
+  settings.legendValueRange = styleConfiguration->hasLegendValueRange;
+  settings.legendLowerRange = styleConfiguration->legendLowerRange;
+  settings.legendUpperRange = styleConfiguration->legendUpperRange;
+  settings.hasNodataValue = dataSource->getDataObject(0)->hasNodataValue;
   width = dataSource->dWidth;
   height = dataSource->dHeight;
-  legendLog = styleConfiguration->legendLog;
-  if (legendLog > 0) {
-    legendLogAsLog = log10(legendLog);
+  settings.legendLog = styleConfiguration->legendLog;
+  if (settings.legendLog > 0) {
+    settings.legendLogAsLog = log10(settings.legendLog);
   } else {
-    legendLogAsLog = 0;
+    settings.legendLogAsLog = 0;
   }
-  legendScale = styleConfiguration->legendScale;
-  legendOffset = styleConfiguration->legendOffset;
+  settings.legendScale = styleConfiguration->legendScale;
+  settings.legendOffset = styleConfiguration->legendOffset;
 
   internalWidth = width;
   internalHeight = height;
@@ -88,22 +88,22 @@ void CAreaMapper::init(CDataSource *dataSource, CDrawImage *drawImage, int tileW
   if (styleConfiguration != NULL && styleConfiguration->styleConfig != NULL && styleConfiguration->styleConfig->RenderSettings.size() == 1) {
     CT::string renderHint = styleConfiguration->styleConfig->RenderSettings[0]->attr.renderhint;
     if (renderHint.equals(RENDERHINT_DISCRETECLASSES)) {
-      isUsingShadeIntervals = true;
+      settings.isUsingShadeIntervals = true;
     }
   }
 
   /* Make a shorthand vector from the shadeInterval configuration*/
-  if (isUsingShadeIntervals) {
+  if (settings.isUsingShadeIntervals) {
     int numShadeDefs = (int)styleConfiguration->shadeIntervals->size();
-    intervals.reserve(numShadeDefs);
+    settings.intervals.reserve(numShadeDefs);
     for (int j = 0; j < numShadeDefs; j++) {
       CServerConfig::XMLE_ShadeInterval *shadeInterVal = ((*styleConfiguration->shadeIntervals)[j]);
-      intervals.push_back(Interval(shadeInterVal->attr.min.toFloat(), shadeInterVal->attr.max.toFloat(), CColor(shadeInterVal->attr.fillcolor.c_str())));
+      settings.intervals.push_back(CDrawFunctionSettings::Interval(shadeInterVal->attr.min.toFloat(), shadeInterVal->attr.max.toFloat(), CColor(shadeInterVal->attr.fillcolor.c_str())));
       /* Check for bgcolor */
       if (j == 0) {
         if (shadeInterVal->attr.bgcolor.empty() == false) {
-          bgColorDefined = true;
-          bgColor = CColor(shadeInterVal->attr.bgcolor.c_str());
+          settings.bgColorDefined = true;
+          settings.bgColor = CColor(shadeInterVal->attr.bgcolor.c_str());
         }
       }
     }
@@ -146,8 +146,8 @@ int CAreaMapper::drawTile(double *x_corners, double *y_corners, int &dDestX, int
 }
 
 template <class T> int CAreaMapper::myDrawRawTile(const T *data, double *x_corners, double *y_corners, int &dDestX, int &dDestY) {
-  int imageWidth = drawImage->getWidth();
-  int imageHeight = drawImage->getHeight();
+  int imageWidth = settings.drawImage->getWidth();
+  int imageHeight = settings.drawImage->getHeight();
   int k;
   for (k = 0; k < 4; k++) {
     if (fabs(x_corners[k] - x_corners[0]) >= fabs(dfSourceBBOX[2] - dfSourceBBOX[0])) break;
@@ -174,7 +174,6 @@ template <class T> int CAreaMapper::myDrawRawTile(const T *data, double *x_corne
 
   bool isNodata = false;
   T val;
-  T nodataValue = (T)dfNodataValue;
   size_t imgpointer;
   double destBoxTL[2] = {x_corners[3], y_corners[3]};
   double destBoxTR[2] = {x_corners[0], y_corners[0]};
@@ -218,47 +217,7 @@ template <class T> int CAreaMapper::myDrawRawTile(const T *data, double *x_corne
       if (sourceSampleX >= 0 && sourceSampleX < width && sourceSampleY >= 0 && sourceSampleY < height && destPixelX >= 0 && destPixelY >= 0 && destPixelX < imageWidth && destPixelY < imageHeight) {
         imgpointer = sourceSampleX + (sourceSampleY)*width;
         val = data[imgpointer];
-
-        /*
-          Please note that this is part of the fast renderer. Changes to this routine should also be implemented in:
-
-          adagucserverEC/CImgWarpNearestNeighbour.h, DrawFunction
-        */
-        isNodata = false;
-        if (hasNodataValue) {
-          if (val == nodataValue) isNodata = true;
-        }
-        if (!(val == val)) isNodata = true;
-        if (!isNodata)
-          if (legendValueRange)
-            if (val < legendLowerRange || val > legendUpperRange) isNodata = true;
-        if (!isNodata) {
-          if (isUsingShadeIntervals) {
-            bool pixelSet = false; // Remember if a pixel was set. If not set and bgColorDefined is defined, draw the background color.
-            for (size_t j = 0; (j < intervals.size() && pixelSet == false); j += 1) {
-              if (val >= intervals[j].min && val < intervals[j].max) {
-                drawImage->setPixel(destPixelX, dstpixel_y + dDestY, intervals[j].color);
-                pixelSet = true;
-              }
-            }
-            if (bgColorDefined && pixelSet == false) {
-              drawImage->setPixel(destPixelX, dstpixel_y + dDestY, bgColor);
-            }
-          } else {
-            if (legendLog != 0) {
-              if (val > 0) {
-                val = (T)(log10(val) / legendLogAsLog);
-              } else
-                val = (T)(-legendOffset);
-            }
-            int pcolorind = (int)(val * legendScale + legendOffset);
-            if (pcolorind >= 239)
-              pcolorind = 239;
-            else if (pcolorind <= 0)
-              pcolorind = 0;
-            drawImage->setPixelIndexed(destPixelX, dstpixel_y + dDestY, pcolorind);
-          }
-        }
+        setPixelInDrawImage(destPixelX, destPixelY, val, &settings);
 
         if (debug) {
           bool draw = false;
@@ -273,10 +232,10 @@ template <class T> int CAreaMapper::myDrawRawTile(const T *data, double *x_corne
             draw2 = true;
           }
           if (draw) {
-            drawImage->setPixelIndexed(destPixelX, destPixelY, 249);
+            settings.drawImage->setPixelIndexed(destPixelX, destPixelY, 249);
           }
           if (draw2) {
-            drawImage->setPixelIndexed(destPixelX, destPixelY, 244);
+            settings.drawImage->setPixelIndexed(destPixelX, destPixelY, 244);
           }
         }
       }
