+++ adagucserverEC/CAreaMapper.h
@@ -29,42 +29,23 @@
 #include "CDebugger.h"
 #include "CDataSource.h"
 #include "CDrawImage.h"
+#include "CDrawFunction.h"
 #include <stdio.h>
 
 class CAreaMapper {
 private:
   DEF_ERRORFUNCTION();
-  bool isUsingShadeIntervals = false;
-  class Interval {
-  public:
-    float min;
-    float max;
-    CColor color;
-    Interval(float min, float max, CColor color) {
-      this->min = min;
-      this->max = max;
-      this->color = color;
-    }
-  };
-  std::vector<Interval> intervals;
-  CColor bgColor;
-  bool bgColorDefined = false;
+
+  CDrawFunctionSettings settings;
+
   double dfTileWidth, dfTileHeight;
   double dfSourceBBOX[4];
   double dfImageBBOX[4];
-  double dfNodataValue;
-  double legendLowerRange;
-  double legendUpperRange;
-  double legendValueRange;
-  double hasNodataValue;
   int width, height;
   int internalWidth, internalHeight;
-  float legendLog, legendScale, legendOffset;
-  float legendLogAsLog;
   CDataSource *dataSource;
-  CDrawImage *drawImage;
   bool debug;
-  ;
+
   template <typename T> int myDrawRawTile(const T *data, double *x_corners, double *y_corners, int &dDestX, int &dDestY);
 
 public:
